pyproject.toml
[project]
name = "wake-ai"
version = "0.1.1"
description = "AI-powered smart contract security analysis framework"
license = { text = "ISC" }
authors = [{ name = "Ackee Blockchain" }]
readme = "README.md"
requires-python = ">=3.10"
keywords = [
    "ethereum",
    "solidity",
    "security",
    "ai",
    "audit",
    "smart-contracts",
    "claude",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: ISC License (ISCL)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "click>=8.0",
    "rich>=13.3.2",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "pydantic>=2.0",
    "eth-wake>=4.20.0",
    "rich-click>=1.8.9",
    "claude-code-sdk>=0.0.20",
]

[project.optional-dependencies]
dev = ["black>=22.0", "isort>=5.0", "mypy>=1.0"]

[project.scripts]
wake-ai = "wake_ai.cli:main"

[project.entry-points."wake.plugins.printers"]
printers = "wake_ai.printers"

[project.entry-points."wake_ai.plugins.workflows"]
workflows = "wake_ai.flows"

[project.urls]
Homepage = "https://github.com/wakehacker/wake-ai"
Repository = "https://github.com/wakehacker/wake-ai"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["wake_ai"]

[tool.black]
target-version = ["py310"]
line-length = 88

[tool.isort]
profile = "black"
skip_gitignore = true
